修改一些提示
修改一些提示

+++ QFForum/QFForum/ZGYForumNet+getComments.m
@@ -21,8 +21,8 @@
     //拼接URL
     NSMutableString * URLStr=[[NSMutableString alloc]init];
     [URLStr appendFormat:@"%@%@?",ZGY_DOMAIN_NAME,ZGY_COMMENT_GET_CS];
-    [URLStr appendFormat:@"&pageNum=%ld",pageNum];
-    [URLStr appendFormat:@"&pageCount=%ld",pageCount];
+    [URLStr appendFormat:@"&pageNum=%ld",(long)pageNum];
+    [URLStr appendFormat:@"&pageCount=%ld",(long)pageCount];
     [URLStr appendFormat:@"&titleId=%@",titleId];
     [URLStr appendFormat:@"&sign=%@",md5Str];
     [URLStr appendFormat:@"&timeStamp=%@",timeStamp];
